develop: removing unnecesary routing, bad domain implementation, bad controller methods
